requirements.txt
transformers
datasets
jiwer
peft
lightning
bitsandbytes>=0.38.1
accelerate
gputil
wandb
lion-pytorch